Refactor: type 분리 및 미사용 type 파일 삭제
