--- src/components/ChartLeft/ChartLeft.types.ts
@@ -1,1 +1,0 @@
-export {};
