--- src/components/ChartRight/ChartRight.types.ts
@@ -1,1 +1,0 @@
-export {};
