--- src/components/ChatBox/ChatBox.types.ts
@@ -1,1 +1,0 @@
-export {};
