--- src/components/SendMessage/SendMessage.types.ts
@@ -1,1 +1,0 @@
-export {};
