--- src/components/Table/Table.types.ts
@@ -1,1 +1,0 @@
-export {};
