--- src/components/common/Footer/Footer.types.ts
@@ -1,1 +1,0 @@
-export {};
