--- src/components/common/Header/Header.types.ts
@@ -1,1 +1,0 @@
-export {};
