--- src/pages/home/Home.types.ts
@@ -1,1 +1,0 @@
-export {};
